requirements.txt

requests==2.31.0
streamlit==1.25.0

langchain==0.0.235
ctransformers==0.2.21

python-docx==0.8.11

Pillow==9.5.0

